+++ VideoGenerator_v1.2.7.0/Services/Configuration/IDialogueStore.cs
@@ -5,4 +5,8 @@ public interface IDialogueStore
     string GetDialogue(string language, string folderName);
 
     void SetDialogue(string language, string folderName, string text);
+
+    bool IsDialogueValidated(string language, string folderName);
+
+    void SetDialogueValidation(string language, string folderName, bool isValidated);
 }